Unit test for repeated reserve
Summary:
Unit test coverage for the "here comes n more" pattern: where the user reserves the existing size, and then some.  This pattern should never result in a *shrinking* reallocation.

Note that F14VectorMap has a bug with respect to this behavior which I will fix in a follow-up diff.

Reviewed By: yfeldblum

Differential Revision: D47407164

fbshipit-source-id: ead373f15ee752ebb19f6d3c0cf3720edb4526fb