Switch to ObserverContainer (8/x): Don't call detachNetworkSocket until move complete
Summary:
This diff is part of a larger change to use `folly::ObserverContainer` (introduced in D27062840) with `AsyncSocket`.

The existing (pseudo) move constructor for `AsyncSocket` calls `detachNetworkSocket` on the old socket to both detach the network socket and get the underlying `NetworkSocket` to set for `fd_` for the new `AsyncSocket`.

However, because the `ObserverContainer` is not moved from the old socket to the new one until later on (given the ordering of the members in the `AsyncSocket` class), observers attached to the old socket are informed about the detach, which in turn creates confusion. Legacy observer implementations had logic to handle this, but we want to avoid the need for such logic with `ObserverContainer`.

To mitigate, this diff changes `AsyncSocket` constructor so that `fd_` is initialized based on the value returned by `oldAsyncSocket->getNetworkSocket()`. `detachNetworkSocket` is then called AFTER completing the move of the `ObserverContainer`. This approach maintains behavior for legacy observers while providing the expected behavior for new observers.

Reviewed By: cristianlumezanu

Differential Revision: D47130538

fbshipit-source-id: 45f5c546c748665788876bbd2330b130c08253a1