build-size tweaks to ThreadLocalPtr helper ElementWrapper::set
Summary:
The overload not taking any deleter does only one thing which is documented as fallible - i.e., maybe throwing -, which is to acquire a lock on a mutex.

But, in practice, lock acquisition failure is a logic error indicating a bug in `ThreadLocalPtr` and not a runtime error that must be handled, so termination should be a reasonable outcome. We mark the function as `noexcept` to force that termination.

Therefore, there is no longer a need in that overload of `ElementWrapper::set` to handle failure, and the failure-handling code may be removed.

Reviewed By: luciang

Differential Revision: D47659980

fbshipit-source-id: 3b47d864c6b41e4b60faf7f4aba72892569f305c