Fix ElfFile::getSectionByIndex() to handle ELFs with bad section header offsets
Summary:
It is possible for the section and program headers in an ELF to not point to valid offsets. There is already a check for this in `ElfFile::iterateSectionHeaders()`:

https://www.internalfb.com/code/fbsource/[f71ad4d91873][history]/fbcode/folly/experimental/symbolizer/Elf-inl.h?lines=46-51

but previously this was not handled when calling `getSectionByIndex()`. This diff adds a check to `getSectionByIndex()` and returns `nullptr` in the invalid offset case.

Reviewed By: hodgesds

Differential Revision: D47882442

fbshipit-source-id: 93850a759461a7d311cbca2e7aaee54c106d8760