CMake - add missing headers dependencies to test targets
Some of the tests targets do not declare all the headers they depend on.
E.g., synchronization/test/BatonTest.cpp depends on BatonTestHelpers.h
but the latter is missing in the `HEADERS`.

Discovered these while attempting to build Folly w/ Bazel (not sure if
this is a good idea). Likewise, I'm not sure what would be the best way
to prevent such, since CMake tolerates such.