Cancel the timeout in timed_wait on success
Summary:
`timed_wait` is not expected to frequently time-out, so by canceling the timer we can save a wake-up of the future timekeeper eventbase.

In reality we don't always save it because in order to cancel the timeout we have to enqueue a task in the eventbase, which also may wake it up, unless the queue is already non-empty in which case we can enjoy some batching, but on average this should be a minor improvement (and we may optimize that separately).

Other timeout primitives like `folly::coro::timeout` already do this.

Reviewed By: iahs

Differential Revision: D48970885

fbshipit-source-id: a9a5711dc2f80e19f8dfc6d3bea837257e2d3b18