Speed up ElfFile::getDefinitionByAddress
Summary:
This diff removes the indirection in std::function from `iterateSymbolsWithTypes`. Given that the process of searching for a symbol by address is a tight loop, the generated code before was substantially worse. Now the inner loop is very tight.

https://pxl.cl/3gTdn

Also captured `address` by value so that it would not be reloaded on each loop iteration.

Reviewed By: Gownta

Differential Revision: D48872564

fbshipit-source-id: d7db8a9aa2e832b61ba59f23e31eefd6494e18a5