Refactor Timekeeper tests
Summary:
We already have 2 `Timekeeper` implementations and I'm working on a third one, currently the test suite of one is copy-pasted from the other and it doesn't even exercise its implementation, since most of the tests rely on the default timekeeper without replacing it.

No tests are added, removed, or substantially changed.

This diff
- Makes the test suite parametrized so we can reuse it
- Splits the tests that don't exercise the default timekeeper into their own non-parametrized suite

Reviewed By: Gownta

Differential Revision: D49012813

fbshipit-source-id: a5eed38f99f7b47327613f90d61a9485f6cfa15c