Implement a heap-based future timekeeper
Summary:
The default future `Timekeeper`, `ThreadWheelTimekeeper`, uses an `EventBase` driven by a dedicated thread and a `HHWheelTimer` to manage the timeouts. For each timeout scheduled, the event base is generally awoken twice: once to schedule the timeout, another to either fulfill it or to process its cancellation.

In practice, the vast majority of timeouts come from `within`/`timed_wait`/`timeout` calls, and are almost always cancelled as the operation succeeds. In these cases, we're wasting the wake-ups, which are expensive.

This diff introduces a new `Timekeeper`, `HeapTimekeeper`, which also uses a dedicated thread but it manages explicitly the set of timeouts. This enables to take some shortcuts:

- When scheduling a timeout, the operation is enqueued and the thread, if sleeping, is awoken only if it would not wake up already before the timeout's expiration. This should be the common case. (The thread is also awoken in the queue gets too long)
- When cancelling a timeout, the promise is cancelled inline, and the operation to delete the timeout from the tracked set is enqueued, but the thread is not explicitly awoken, as there is no urgency in processing the deletion, since doing that would not avoid any wake-up.

Instead of the hierarchical hashed wheel used by `HHWheelTimer` we simply adopt an intrusive heap. While the former has roughly constant time operations, the cost of maintaining the set of timeouts is so negligible that it doesn't warrant the additional complexity; as the insert/delete operations are processed in batches, some of the cache misses from the top levels of the heap are amortized anyway. Furthermore, this removes the need to quantize the timeouts into ticks, improving the accuracy of the timekeeper.

The stress test added in D49015598 gives an idea of the accuracy:

```
$ buck2 run mode/opt folly/futures/test:thread_wheel_timekeeper_test -- --gtest_filter=*Stress
...
I0908 08:31:36.273962 4184654 TimekeeperTestLib.h:556] Successful completions: 785, avg delay: 1129 us, max delay: 3064 us

$ buck2 run mode/opt folly/experimental/test:time -- --gtest_filter=*Stress
...
I0908 08:32:21.399613 4186827 TimekeeperTestLib.h:556] Successful completions: 801, avg delay: 135 us, max delay: 465 us

$ buck2 run mode/opt folly/futures/test:heap_timekeeper_test -- --gtest_filter=*Stress*
...
I0908 08:32:31.819116 4188374 TimekeeperTestLib.h:556] Successful completions: 809, avg delay: 76 us, max delay: 169 us
```

We can see that for the `HHWheelTimer`-based timekeepers the accuracy is limited by the tick interval (1ms for default, 200us for high-res), while `HeapTimekeeper` is mostly limited by how fast is the kernel at waking up the worker thread, so the accuracy is much better than both.

Reviewed By: dmm-fb

Differential Revision: D49060468

fbshipit-source-id: e1eac770e530e7ec757484dd62bf55bee138d244