Add epoll based backend
Summary:
This is an epoll based EventBase backend (derived from EventBaseBackendBase).
It has the same functionality as the libevent base one with some implementation differences.
The timers are managed by a timer fd - similar to the high res one available in newer versions of libevent.
We expose the epollfd to allow for monitoring by other epoll instances.

Reviewed By: ot

Differential Revision: D42281517

fbshipit-source-id: 7b097aa8d06bffc8b3f622c77e74a814dfb17570