Mockable / fakeable clock
Summary:
This diff introduces `folly::ClockInterface`, which makes it easy to build a mockable / fakeable clock.

Being able to control the times returned from clocks is often useful in tests and emulation. I've ended up building a number of clock abstraction layers over the past few years to support such use cases, and have found if you build this support in from the beginning, it saves you down the road.

The most common interfaces and implementations are predefined:
- `folly::SteadyClock`: interface for a steady clock based on `std::chrono::steady_clock`
- `folly::SteadyClockImpl`: implementation of a steady clock based on `std::chrono::steady_clock`
- `folly::SystemClock`: interface for a system clock based on `std::chrono::system_clock`
- `folly::SystemClockImpl`: implementation of a system clock based on `std::chrono::system_clock`

Reviewed By: metaduc

Differential Revision: D48688441

fbshipit-source-id: 75fb2fa5da46b0246257e0cc3251f3950795264e