Add a strict mode for EventBase::isInEventBaseThread()
Summary:
If an `EventBase` is managed by an executor that doesn't run the loop continuously, it is not safe to use `isInEventBaseThread()` because it can race with the starting and stopping on the loop. In this case, `isInEventBaseThread()` should always return `false` if the loop is not running, otherwise it is possible to run non-thread-safe logic assuming that the current thread is the event base thread, concurrently with another thread that starts running the loop.

We can add a strict mode for this case, where `isInEventBaseThread()` always returns false if the loop is not running, and thus it can only return true if called from the loop thread while the loop is running.

Reviewed By: dmm-fb

Differential Revision: D49658388

fbshipit-source-id: 266f5c794396352e64bf84a74cb394e91e4da4ce