Skip setting request context in coro sleep() and timed_wait() timeouts
Summary:
We spend a nontrivial amount of cycles setting the request context in the `FutureTimekeepr` thread, almost all coming from the `coro::sleep` callback, which just posts a baton. This kind of plumbing is usually not covered by the request context, because it does not include user logic and the cost of the operation can be comparable to the cost of setting the request context.

This diff does not propagate the current request context to the callback for `sleep()`, and while at it for `timed_wait()` as well.

Reviewed By: Gownta

Differential Revision: D49578271

fbshipit-source-id: f374da8b7565cc437604f417d7ce0c146fd6434a