Optimize construction of most {Future,Promise}Exceptions
Summary:
Some exception such as `FutureCancellation` and `BrokenPromise` are constructed at high volume because they are used for non-exceptional paths. The base class `logic_error` always copies its argument into an internal `std::string`, and most of the error messages do not fit in inline storage.

Since the error messages are all static, we can specialize the `what()` to avoid the copy.

Reviewed By: ispeters

Differential Revision: D49638973

fbshipit-source-id: 3aa91f1655d10e5eda28c70f24745eefd92bde0b