revise Optional and Expected coroutine return-object tracking
Summary:
LLVM 15 built code crashes n the existing coroutine implementation for `folly::Expected`.

LLVM 15 has changed how it handles the coroutine return-object in https://reviews.llvm.org/D117087, in a way which is incompatible with the existing code for `folly::Expected`. It looks like the new behavior is the same eager-conversion alternative that MSVC 2019 update 16.5 finally retired, so the new behavior might be classified as a regression.

Fix this regression by revising how Optional and Expected manage the coroutine promise return-object and by detecting which behavior is required. Lift the behavior-detection facility to the base coroutine header.

We have a patched LLVM 15 with the offending upstream patch reverted so, under LLVM, the detection must be more complex than a simple version check; we do the detection at runtime since, conveniently, LLVM manages to optimize all of the runtime detection down to a constant. Other alternatives we considered are: (1) to return to the upstream behavior and do a straightforward version check or (2) to have the patched compiler, or builds using the patched compiler, predefine a preprocessor symbol and have the detection look for that symbol.

Reviewed By: bcardosolopes

Differential Revision: D42260201

fbshipit-source-id: cbef347e373c8580b3ad53b1c78c6bab8e8a7db6