Refactor co_scope_exit to prepare it for using in AsyncGenerator
Summary: The diff delays setting context parameters of `co_scope_exit` (coroutine continuation, async stack frame, executor), that allows to execute the `co_scope_exit` task in a context different from the context where the task was attached. This is needed to add support for `co_scope_exit` in `AsyncGenerator`.

Reviewed By: iahs

Differential Revision: D49652557

fbshipit-source-id: a287ed7015cfff7aae591475462286217fd4aa03