Fix an MSVC signed/unsigned comparison in EventBaseBackendBase
Summary:
```
xplat\folly\io\async\EventBaseBackendBase.cpp(134): warning C4018: '>=': signed/unsigned mismatch
```
`h->name` is unsigned while `msg_namelen` is signed, so this comparison has always been treated as unsigned.

I guess `socklen_t msg_namelen` could be `-1` if there is an error, that would treat it as a large number and evaluate the expression, choosing the branch:
```
    out.name = total.subpiece(sizeof(H), h->name);
```
Suppress the warning if there is no interest in changing this.

Reviewed By: Gownta

Differential Revision: D50141494

fbshipit-source-id: 2f88da2c911204873142ba65a0637ed2f3255753