Support co_scope_exit in AsyncGenerator
Summary: Allow to attack `co_scope_exit` tasks to `AsyncGenerator`. The tasks are executed after destruction of `AsyncGenerator` coroutine frame when `cleanup()` is `co_await`ed. To not break the contract of existing `AsyncGenerator` the functionality is enabled only when a new template parameter `RequiresCleanup` is set to `true`. For convenience there is also a `CleanableAsyncGenerator` alias for `AsyncGenerator` with `destructible` set to true.

Reviewed By: capickett

Differential Revision: D49734568

fbshipit-source-id: 332972a66c7fa6cc7ba08951ba964bee326b2643