Fix shared_lock<SharedMutex>::release()
Summary:
Tokenful `lock_shared()` must be paired with a tokenful `unlock_shared()`, because tokenfulness is recorded in the deferred reader value, so a tokenless `unlock_shared()` would fail to clear the deferred reader slot that was acquired by a tokenless `lock_shared()`.

This was not clearly documented (it now is with this diff).

In particular, this makes `std::shared_lock::release()` incorrect in the specialization, since it makes it impossible to unlock the mutex as the token is lost.

To support this case, this diff introduces a method that consumes the token without unlocking the mutex, so that the mutex can later be unlocked with a tokenless `unlock_shared()`.

Reviewed By: ilvokhin

Differential Revision: D50645432

fbshipit-source-id: 806e50df879ba6494be21c71133778c79651f14a