Prevent double deletion in DetachedBarrierTask with MSVC
Summary:
This seems to be a currently unaddressed bug in the MSVC compiler.
It manifests itself as a crash (with no stack trace) during the clean-up phase of tests.

Let's look at this link first: https://stackoverflow.com/questions/71106923/calling-destroy-from-final-suspend-results-in-a-crash

Here, the example code runs differently between Clang/GCC and MSVC and we observe three calls to the destructor compared to the expected two.
See diffs D50952075 and D50972393 for two different ways to mitigate the crash. D50952075 explores the approach used in this diff and gates the call to the `coroutine_handle::destroy` when we detect MSVC. On the other hand, D50972393 adds the `/Zc:nrvo-` flag to MSVC to disable named return value optimization. This did somehow suppress the crash but we still see three prints at the output.

At [another place](https://developercommunity.visualstudio.com/t/c20-coroutine-double-destroy-coroutine/1331004), a user reported a similar issue where the workaround of disabling NRVO was brought up. Currently, the MSVC team says that the fix is ready but there is no release with the fix at the moment. Technically, older releases of MSVC will still be affected by this issue.

Differential Revision: D50950293

fbshipit-source-id: 66936ea7fb8b8cf0a724eb559bd60206de4a3656