Remove extraneous semicolon from folly/synchronization/HazptrHolder.h
Summary:
Extraneous semicolons are a code smell and can mask more serious problems. `-Wextra-semi-stmt` finds them.

This diff removes an extraneous semicolon or adjusts a macro so that a semicolon is required after the macro (making it look like a standard function).

This file is drawn from a heavy-hitting list, so fixing this problem will allow *many* other files to take advantage of the safety `-Wextra-semi-stmt` offers.

This should be a low-risk diff: if it compiles, it works.

Reviewed By: Orvid

Differential Revision: D51631113

fbshipit-source-id: cb26e946c075a9b45d11a6c1d47a6b523188c407