fix rare SerialExecutor heap leak
Summary:
SerialExecutor can cause heap leaks when the parent Executor discards the SerialExecutor::worker() tasks instead of executing them: the worker() is responsible for dequeuing the tasks passed to SerialExecutor::add(), and if the parent Executor just destructs the SerialExecutor::worker() task instead of executing it (which for example happens if that executor was a weakRef whose underlying Executor already got destructed) then SerialExecutor enters a broken state, with non-empty task queue, and worker() never running again (because scheduled_ never can flip from 0 to 1 again). The ParentExecutorDiscardsFunc test added in this diff here demonstrates & fixes the problem.

BTW the code is in this diff was all authored by ot, all I did was deliver the initial bug repro, and an early draft for a fix.

Reviewed By: ot

Differential Revision: D51908477

fbshipit-source-id: 8213d792a041c3a28c45b3c99adb0d408befe29b