ThreadPoolExecutorTest: ensure SingleThreadedCPUThreadPoolExecutor is stopped before running the base class dtor
Summary:
A recent change to googletest (D51155941) exposed this data race.

The race is on the vptr - while the dtor runs the executor thread is still running and making calls to the virtual `keepAliveAcquire()` method.

To fix, simply make sure the executor is stopped before the dtor finishes.

NOTE: this only affects a test.

Reviewed By: ot

Differential Revision: D52155086

fbshipit-source-id: 5749cdb90c8a2aab35a74a89d0e0bed6b9c61752