IOThreadPoolExecutorBase::IOObserver
Summary:
Several observers extract the `EventBase` pointer from the `Thread` handle, but this only works under the assumption that threads and evbs are 1:1, which only holds for `IOThreadPoolExecutor`.

All these observers actually only require to track the evbs used in the executor, so we can add a dedicated observer to `IOThreadPoolExecutorBase` for that purpose, that works correctly even if evbs and threads are independent.

Reviewed By: dmm-fb

Differential Revision: D52123131

fbshipit-source-id: 8639b756c6dcb0c856fb19f5b36a444e74f90382