Fix MuxIOThreadPoolExecutor shutdown behavior
Summary:
The `wakeup()` function is broken: there is no correspondence between threads and EventBases, so we may unset `shouldRun` on one thread and then wake up another.
Furthermore, the destructor is stopping the threads before signaling the main thread, so if the eventbase we pick to wake up is disarmed, on its way to be handled, no thread can do that and the main loop is stuck.

This diff fixes thread stopping by relying on poison tasks (similarly to `CPUThreadPoolExecutor`), and wakes up the main thread using the eventfd.

We can also remove the trick of enqueueing a callback when constructing the EventBase to set up the internal notification queue: if we hold a keepalive to it, like `loopForever()` does, the notification queue is set up as a non-internal event, so libevent waits for it correctly.

Reviewed By: dmm-fb

Differential Revision: D52180081

fbshipit-source-id: 8d5fed185a5238c4eb43146c4198abdd9421c205