Revert arm support in AtomicIntrusiveLinkedList
Summary:
D51609008 added `arm()` support to `AtomicIntrusiveLinkedList`, but I believe that this functionality was a poor fit for the class:

- Using any traversal method on an armed list would be undefined behavior. Even if we wanted to fix that, would would be the correct behavior?
- The semantics are tricky and hard to document
- I don't think we'll have more than one use case for this

Given that `arm()` and `insert()` are just a handful of lines of code each, I wouldn't try to over-generalize this and just implement them directly inside `MuxIOThreadPoolExecutor`.

`AtomicIntrusiveLinkedList` is reverted to its pre-D51609008 state.

Reviewed By: dmm-fb

Differential Revision: D52141219

fbshipit-source-id: 15f2d3c2ff8f3f51f0f161e660632eab22837bd5