symbolizer: add support for looking up CUs using the .gdb_index if present
This provides similar information as `.debug_aranges` but is better supported on DWARFv4 and with clang. Since it is generated at link time, it has the additional benefit that it is complete, meaning it can return negative results (for symbols such as `_start` that may be implemented in assembly and thus have no associated debug info) without walking the full debuginfo list.