Revert D51998580: Multisect successfully blamed "D51998580: [folly] Improve large-buffer crc32c performance by 50-85%" for otest failure
Summary:
This diff is reverting D51998580
D51998580: [folly] Improve large-buffer crc32c performance by 50-85% by conradmeyer has been identified to be causing the following test failure:

Tests affected:
- [horizon/platform/download_service:lib-unittest - handler::handler::test::test_handle_hedwig_complete_success](https://www.internalfb.com/intern/test/562950049273471/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3913363
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D52647553

fbshipit-source-id: 2ca7b761dcdedf2ef5298a9585e22decb32dd00d