Spin for longer time (cycles) in `SharedMutex::waitForZeroBits`
Summary:
Motivational hypothesis for this change is previous attempt to tune spin
count gave good results, because latency of the `pause` instruction changed from
~10 to 140+ cycles on Skylake microarchitecture, but original code was tuned for
pred-Skylake CPUs. Moreover, win from the previous change might come
from single huge user from internal workload, which likely should be analysed
separately as it has unusually big critical section.

This change switched from count based spin to time (cycles) base spin to spin
approximately the same amount of time regardless of microarchitecture.

Constant of cycles to spin for was chosen to be equal ~2 us on 2 GHz CPU:
4000 cycles = 2000 ns = 2 us @ 2 GHz.

Spin time was increased, comparing to previous value (~78.38 ns, measured with
benchmarks) based on the observed usage pattern for `folly::Synchronized`:
majority of usages have small critical section and short amount of spinning
should be beneficial for them, but not so short as it was before to not sleep
on futex.

Reviewed By: ot

Differential Revision: D52541906

fbshipit-source-id: 8a3dd9120dc8bec682aac8af440769a9c45dc127