fix TSAN-reported data-race in AsyncSocketTest.QueueTimeout
Summary:
The data-race is between the two uses of stack-allocated variable `kEventLoopTime`:
* In the current acceptor thread, where it is used as an argument to `sleep_for`.
* In the callback passed to the main thread, where it is captured by-reference and used as an argument to `tryRunAfterDelay`.

The `sleep_for` prevents this data-race from being detected by ASAN as a stack-use-after-scope since, in practice, the variable lives long enough to be used only while it is in scope.

There are several straightforward fixes for this data-race, of which one was chosen but any could have been chosen:
* Since `kEventLoopTime` is compile-time constant, it may be marked `static constexpr` rather than `constexpr` to move its allocation off of the stack.
* Capture `kEventLoopTime` by-value rather than by-reference into the callback so that all uses of the stack-allocated variable occur in the current acceptor thread.
* Wait for the callback passed to the main thread to be run, such as by switching from `runInEventBaseThread` to `runInEventBaseThreadAndWait`.

The data-race which TSAN detects and reports is cryptic - it opaquely points to the call to `sleep_for` as a false synchronization, and it points to an internal stack-allocated buffer in `fmt::format` as the race location. This has led multiple investigators down wrong paths, including one hypothesis that there is some data-race within `fmt::format`.

Differential Revision: D52803848

fbshipit-source-id: 0ac7104e94c7ff83875f851928182902644ab07f