Increase compiler warnings in folly/PACKAGE
Summary:
LLVM-15 has many useful compiler warnings that are not yet enabled globally. In order to enable them globally we are first enabling them on a per-directory basis.

This diff enables additional warning flags for this directory. If the builds succeed, the diff should be safe to land.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

(1 file modified.)

Reviewed By: dmm-fb

Differential Revision: D52931439

fbshipit-source-id: 3019692f9fef96202784353aca05f998fcb45eb4