Slight improvement to error message in folly/Benchmark.cpp
Summary:
I was reading the flags of this file in a help message and I noticed that the help message for `--bm_json_verbose` refers to `--json` and `--bm_relative_to` as "above" flags. However, it comes before those 2 flags in alphabetical order so in the help message they actually appear below the flag value. This slightly confused me so I just removed the "above" phrasing.

Feel free to reject this diff if you don't think this is a good change.

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: Gownta

Differential Revision: D47979516

fbshipit-source-id: bc0c8cf4e47887b2c68766ddbd71166a57c7ea5a