Fix alignment of small_vector to match clang/GCC for MSVC
Summary:
With `_MSC_VER` (which clang also defines on windows), `small_vector` is packing the entire class rather than just the internal storage (as is done on GCC/clang on non-Windows platforms).

What this means is that on Windows, alignof(small_vector) is always 1, and this breaks alignment requirements for the stored type, leading to UB. On GCC/clang, alignof(small_vector) is correct and no issues are hit.

This diff changes the application of the MSVC macros to match the use of GCC/clang attributes and only applies them to the internal structs.

Reviewed By: yfeldblum

Differential Revision: D53044143

fbshipit-source-id: 88000a2a079dc5a083291402ffca1108581e978f