Add wrapper for F14HashToken
Summary:
This adds a new type, `F14HashedKey` which is essentially a key-hash pair. I believe this is more robust than the current approach of managing `F14HashToken` completely separately from the key; the `F14HashedKey` maintains the guarantee that the token and key always match.

This adds new methods to `F14FastMap` that takes `F14HashedKey` parameters.

Reviewed By: Gownta

Differential Revision: D53180398

fbshipit-source-id: bc20921b6ca2a7be58af43be81bed3af17991560