Disallow EventBase::terminateLoopSoon() in strict loop thread mode
Summary:
`terminateLoopSoon()` should only be called when owning the `EventBase`; it should not be allowed on an `EventBase` owned by an executor, because it makes it impossible for the executor to know when the `EventBase` is fully drained (callbacks may call it and stop the drain).

This diff disallows it in strict mode (currently only used by `MuxIOThreadPoolExecutor`), and I'll try enabling this in other executors separately.

Reviewed By: Gownta

Differential Revision: D53385601

fbshipit-source-id: 972cdb21cef88e24d6c8ae80b4a1c4621f1658c5