Only define the std::hash<folly::Optional<T>> specialization if std::hash<T> is defined and usable
Summary:
Currently, `folly::Optional` defines the `std::hash<folly::Optional<T>>` specialization even if `std::hash<T>` isn't defined or is otherwise unusable.

This is a problem because it makes it more difficult to implement customized hashing that relies on `std::hash<T>` specializations existing or not existing using SFINAE tricks.

To fix it, I'm stealing some code from libc++'s `std::optional` implementation:
- https://github.com/llvm-mirror/libcxx/blob/master/include/optional,
- https://github.com/llvm-mirror/libcxx/blob/master/include/utility.

I customized this code and put it in `folly/Traits.h` for ease of reuse in case we need it for something else.

Reviewed By: Gownta

Differential Revision: D53822800

fbshipit-source-id: afd0104543a320adac3203dcd1b89062951d3dfd