Tweak EventBasePoller io_uring params
Summary:
Enable some flags that should make things a bit faster in the common case.

For `IORING_SETUP_SINGLE_ISSUER` the ring needs to be created in the same thread that will use it, so I needed to change the setup a bit, instead of creating it in the constructor.

Reviewed By: dmm-fb

Differential Revision: D54074991

fbshipit-source-id: 48445d87daf2e28cc975a04cd915f7037b347cee