fix windows vs2019 compilation with non-constexpr offsetof
Summary:
This line was breaking vs2019 builds which are being used by Horizon's clang-tidy run locally on Windows.

Root cause appears to be using offsetof in a constexpr, which is correct and fine ALMOST everywhere... but Windows for a long time used an illegal offsetof which is non-constexpr.

Fix is just remove the constexpr.

Reviewed By: chadaustin, Orvid, dmpolukhin

Differential Revision: D53553082

fbshipit-source-id: ad79b65c00838431a1ded8df0f60f328993d4e5f